TAT/TaskRecorder/requirements.txt
pyautogui==0.9.54
pynput==1.7.6
Pillow==10.2.0
opencv-python==4.9.0.80
numpy>=1.24.3
pyobjc-framework-Quartz==10.1; sys_platform == 'darwin'
pyobjc-framework-ApplicationServices==10.1; sys_platform == 'darwin'
